+++ examples/kong-dp-local/main.tf
@@ -4,21 +4,18 @@ module "kong" {
   # version = "1.0.0" // first commit
 
   kong-image-tag             = var.kong-image-tag
-  cluster-control-plane      = var.cluster-control-plane
-  cluster-telemetry-endpoint = var.cluster-telemetry-endpoint
   kong-cluster-cert-path     = var.kong-cluster-cert-path
   kong-namespace             = var.kong-namespace
-  service-monitor-enable     = var.service-monitor-enable
-  service-monitor-namespace  = var.service-monitor-namespace
+  kong-values-file           = "${path.module}/data-plane-values.yaml"
 }
 
 /*
 data "aws_eks_cluster" "kong-eks-cluster" {
-  name = "testantoine"
+  name = "$cluster_name"
 }
 
 data "aws_eks_cluster_auth" "kong-eks-cluster" {
-  name = "testantoine"
+  name = "$cluster_name"
 }
 
 provider "aws" {