+++ CHANGELOG.md
@@ -22,3 +22,16 @@
 - New examples using registry
 
 ### Fixed
+
+
+## [0.0.3] - 2023-06-08
+
+### Added
+
+- Add filename variable
+
+### Changed
+
+- Add filename variable
+
+### Fixed